+++ .gitignore
@@ -8,4 +8,6 @@
 node_modules/
 server/node_modules/
 
+public/bundle.js
+
 ngrok
