+++ server/views/pages/game.ejs
@@ -85,7 +85,10 @@
             <% } %>
         </div>
     </div>
-    <script src="../public/main.js"></script>
+<!--    <script type="module" src="../poker-logic/lib/table-state.js"></script>-->
+<!--    <script type="module" src="../server/server-logic.js"></script>-->
+<!--    <script type="module" src="../public/main.js"></script>-->
+    <script src="../public/bundle.js"></script>
 </body>
 
 </html>
